tweaked factorial and fibonacci functions in concurrency test

+++ src/test/scala/objektwerks/concurrency/ConcurrencyTest.scala
@@ -13,15 +13,15 @@ import scala.math.*
 import scala.util.{Failure, Success, Try, Using}
 
 final class FactorialTask(n: Int) extends Callable[Long]:
-  @tailrec def factorial(n: Long, acc: Long = 1): Long = n match
+  @tailrec def factorial(n: Int, acc: Long = 1): Long = n match
     case i if i < 1 => acc
     case _ => factorial(n - 1, acc * n)
 
   def call(): Long = factorial(n)
 
 final class FibonacciTask(n: Int) extends Callable[Long]:
-  def fibonacci(n: Long): Long =
-    @tailrec def loop(n: Long, a: Long, b: Long): Long = n match
+  def fibonacci(n: Int): Long =
+    @tailrec def loop(n: Int, a: Long, b: Long): Long = n match
       case 0 => a
       case _ => loop(n - 1, b, a + b)
     loop(n, 0, 1)
